simplified code, weather is displaying
